[Edit] onEnableの整理
