+++ src/main/java/ga/ganma/minigames/mission.java
@@ -2,6 +2,7 @@ package ga.ganma.minigames;
 
 import org.bukkit.*;
 
+import ga.ganma.minigames.listeners.GameListeners;
 import ga.ganma.minigames.missiontime.*;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
@@ -37,7 +38,7 @@ public class Mission {
 			p.sendTitle(ChatColor.RED + "ミッションが発動された...", "ゲーム時間をループ状態から戻せ！", 10, 60, 20);
 			p.playSound(p.getLocation(), Sound.UI_TOAST_IN, 1, 2);
 		}
-		EventGet.missionS = "ゲームエラーのため、ゲーム時間が1200秒から600秒の間でループするようになった！" + "これを直すために、ある地点にゲームエラーを修復するための岩盤を設置した。"
+		GameListeners.missionS = "ゲームエラーのため、ゲーム時間が1200秒から600秒の間でループするようになった！" + "これを直すために、ある地点にゲームエラーを修復するための岩盤を設置した。"
 				+ "そのボタンを押し、ゲームエラーを直せ！";
 		missionLocation.setX(missionLocation.getX() - 1);
 		TosoCommand.world.getBlockAt(missionLocation).setType(Material.BEDROCK);
@@ -52,7 +53,7 @@ public class Mission {
 			p.sendTitle(ChatColor.RED + "ミッションが発動された...", "プレイヤーの発光を阻止せよ", 10, 60, 20);
 			p.playSound(p.getLocation(), Sound.UI_TOAST_IN, 1, 2);
 		}
-		EventGet.missionS = "今から10分後にプレイヤーが30秒間発光してしまう！" + "しかし、どこかにプレイヤーの発光を阻止するブロックを置いた。"
+		GameListeners.missionS = "今から10分後にプレイヤーが30秒間発光してしまう！" + "しかし、どこかにプレイヤーの発光を阻止するブロックを置いた。"
 				+ "その場所を見つけ出し、その場所にあるアイテムを拾い、プレイヤーの発光を阻止せよ！" + "なお、そのアイテムを拾うと、使った者に対し逃走に有利な条件が付く！";
 		missionLocation.setX(missionLocation.getX() - 1);
 		TosoCommand.world.getBlockAt(missionLocation).setType(Material.REDSTONE_BLOCK);
@@ -72,7 +73,7 @@ public class Mission {
 			p.getInventory().setItem(0, item2);
 			mission2B.put(p, false);
 		}
-		EventGet.missionS = "プレイヤー全員に鍵を渡した。今から10分後に時限装置が鳴り響くようになり、ハンターが集まってくる！"
+		GameListeners.missionS = "プレイヤー全員に鍵を渡した。今から10分後に時限装置が鳴り響くようになり、ハンターが集まってくる！"
 				+ "解除するには、他のプレイヤーが持っている鍵を右クリックして貰う必要がある！" + "なお、鍵は何回でも使用可能である。";
 	}
 
@@ -86,9 +87,9 @@ public class Mission {
 			p.playSound(p.getLocation(), Sound.UI_TOAST_IN, 1, 2);
 		}
 		blockL = missionLocation;
-		EventGet.missionS = "ある地点にクリックすると賞金単価を1秒300円にするブロックを設置した。" + "しかし、そのブロックをクリックするとハンターが20分間強化されてしまう！"
+		GameListeners.missionS = "ある地点にクリックすると賞金単価を1秒300円にするブロックを設置した。" + "しかし、そのブロックをクリックするとハンターが20分間強化されてしまう！"
 				+ "押さなくても何もペナルティは無い。押すか押さないかはプレイヤーの自由だ！";
-		EventGet.chat = false;
+		GameListeners.chat = false;
 		missionLocation.setX(missionLocation.getX() - 1);
 		TosoCommand.world.getBlockAt(missionLocation).setType(Material.DIAMOND_BLOCK);
 	}
