+++ src/main/java/ga/ganma/minigames/missiontime/mission4time.java
@@ -1,9 +1,9 @@
 package ga.ganma.minigames.missiontime;
 
-import ga.ganma.minigames.EventGet;
 import ga.ganma.minigames.Minigames;
 import ga.ganma.minigames.Mission;
 import ga.ganma.minigames.TosoCommand;
+import ga.ganma.minigames.listeners.GameListeners;
 
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -19,7 +19,7 @@ public class Mission4Time extends BukkitRunnable {
 		if (Minigames.gameTime == Mission.mission4t) {
 			Mission.isMission = false;
 			this.cancel();
-			EventGet.chat = true;
+			GameListeners.chat = true;
 			for (Player allp : Bukkit.getServer().getOnlinePlayers()) {
 				allp.sendTitle(ChatColor.RED + "誰も押さなかったため、賞金単価は上がらなかった...", "", 0, 100, 0);
 				allp.playSound(allp.getLocation(), Sound.UI_BUTTON_CLICK, 1, 2);
@@ -27,7 +27,7 @@ public class Mission4Time extends BukkitRunnable {
 			TosoCommand.world.getBlockAt(Mission.blockL).setType(Material.AIR);
 		} else if (!Mission.isMission) {
 			this.cancel();
-			EventGet.chat = true;
+			GameListeners.chat = true;
 			for (Player allp : Bukkit.getServer().getOnlinePlayers()) {
 				allp.sendTitle(ChatColor.RED + "ハンター強化！", Mission.CLEARp.getName() + "により、単価が300円に上がったがハンターが強化された！", 0,
 						100, 0);
