+++ src/main/java/ga/ganma/minigames/tosoCommand.java
@@ -24,29 +24,29 @@ import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.inventory.meta.LeatherArmorMeta;
 
 public class TosoCommand implements CommandExecutor {
-	static Player p;
-	static ItemStack menu1 = new ItemStack(Material.IRON_FENCE);
-	static ItemMeta menu1data = menu1.getItemMeta();
-	static ItemStack menu2 = new ItemStack(Material.STICK);
-	static ItemMeta menu2data = menu2.getItemMeta();
-	static ItemStack menu3 = new ItemStack(Material.IRON_BLOCK);
-	static ItemMeta menu3data = menu3.getItemMeta();
-	static ItemStack menu4 = new ItemStack(Material.REDSTONE_TORCH_ON);
-	static ItemMeta menu4data = menu4.getItemMeta();
-	static ItemStack menu5 = new ItemStack(Material.PAPER);
-	static ItemMeta menu5data = menu5.getItemMeta();
-	static ItemStack menu6 = new ItemStack(Material.PAPER);
-	static ItemMeta menu6data = menu6.getItemMeta();
-	static ItemStack menu7 = new ItemStack(Material.PAPER);
-	static ItemMeta menu7data = menu7.getItemMeta();
-	static ItemStack menu8 = new ItemStack(Material.PAPER);
-	static ItemMeta menu8data = menu8.getItemMeta();
-	static ItemStack close = new ItemStack(Material.REDSTONE_BLOCK);
-	static ItemMeta closedata = close.getItemMeta();
-	static ItemStack hunter1 = new ItemStack((Material.LEATHER_CHESTPLATE));
-	static ItemStack hunter2 = new ItemStack((Material.LEATHER_HELMET));
-	static ItemStack hunter3 = new ItemStack((Material.LEATHER_LEGGINGS));
-	static ItemStack hunter4 = new ItemStack((Material.LEATHER_BOOTS));
+	public static Player p;
+	public static ItemStack menu1 = new ItemStack(Material.IRON_FENCE);
+	public static ItemMeta menu1data = menu1.getItemMeta();
+	public static ItemStack menu2 = new ItemStack(Material.STICK);
+	public static ItemMeta menu2data = menu2.getItemMeta();
+	public static ItemStack menu3 = new ItemStack(Material.IRON_BLOCK);
+	public static ItemMeta menu3data = menu3.getItemMeta();
+	public static ItemStack menu4 = new ItemStack(Material.REDSTONE_TORCH_ON);
+	public static ItemMeta menu4data = menu4.getItemMeta();
+	public static ItemStack menu5 = new ItemStack(Material.PAPER);
+	public static ItemMeta menu5data = menu5.getItemMeta();
+	public static ItemStack menu6 = new ItemStack(Material.PAPER);
+	public static ItemMeta menu6data = menu6.getItemMeta();
+	public static ItemStack menu7 = new ItemStack(Material.PAPER);
+	public static ItemMeta menu7data = menu7.getItemMeta();
+	public static ItemStack menu8 = new ItemStack(Material.PAPER);
+	public static ItemMeta menu8data = menu8.getItemMeta();
+	public static ItemStack close = new ItemStack(Material.REDSTONE_BLOCK);
+	public static ItemMeta closedata = close.getItemMeta();
+	public static ItemStack hunter1 = new ItemStack((Material.LEATHER_CHESTPLATE));
+	public static ItemStack hunter2 = new ItemStack((Material.LEATHER_HELMET));
+	public static ItemStack hunter3 = new ItemStack((Material.LEATHER_LEGGINGS));
+	public static ItemStack hunter4 = new ItemStack((Material.LEATHER_BOOTS));
 	LeatherArmorMeta meta1 = (LeatherArmorMeta) hunter1.getItemMeta();
 	LeatherArmorMeta meta2 = (LeatherArmorMeta) hunter2.getItemMeta();
 	LeatherArmorMeta meta3 = (LeatherArmorMeta) hunter3.getItemMeta();
